added new programs

+++ src/JAVA_DSA/CLASS/DSA_MODULE_1/DSA_Arrays_Carry_Forward_Subarrays/Pick_from_both_sides.java
@@ -1,26 +1,27 @@
 package CLASS.DSA_MODULE_1.DSA_Arrays_Carry_Forward_Subarrays;
 
 public class Pick_from_both_sides {
-    public static long solve(int[]A,int B){
+    public static int solve(int[] A, int B) {
         int n = A.length;
         long sum = 0;
-        for(int i = 0 ; i < B; i ++){
-            sum+=A[i];
+        for (int i = 0; i < B; i++) {
+            sum += A[i];
         }
         long ans = sum;
-        int i = B-1;
-        int j = n-1;
-        while(i >= 0){
+        int i = B - 1;
+        int j = n - 1;
+        while (i >= 0){
             sum-=A[i];
             sum+=A[j];
-            ans = Math.max(ans,sum);
+            ans= Math.max(ans,sum);
             i--;j--;
         }
-        return ans;
+        return (int)ans;
     }
-    public static void main(String[] args){
-        int[]A = {5, -2, 3 , 1, 2};
+
+    public static void main(String[] args) {
+        int[] A = {5, -2, 3, 1, 2};
         int B = 3;
-        System.out.println(solve(A,B));
+        System.out.println(solve(A, B));
     }
 }
